chore: separate class Looper
